# week7/flask-server/openapi_server/models/book_fields.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class BookFields(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, title=None, author=None, year=None):  # noqa: E501
        """BookFields - a model defined in OpenAPI

        :param title: The title of this BookFields.  # noqa: E501
        :type title: str
        :param author: The author of this BookFields.  # noqa: E501
        :type author: str
        :param year: The year of this BookFields.  # noqa: E501
        :type year: int
        """
        self.openapi_types = {
            'title': str,
            'author': str,
            'year': int
        }

        self.attribute_map = {
            'title': 'title',
            'author': 'author',
            'year': 'year'
        }

        self._title = title
        self._author = author
        self._year = year

    @classmethod
    def from_dict(cls, dikt) -> 'BookFields':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BookFields of this BookFields.  # noqa: E501
        :rtype: BookFields
        """
        return util.deserialize_model(dikt, cls)

    @property
    def title(self) -> str:
        """Gets the title of this BookFields.


        :return: The title of this BookFields.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this BookFields.


        :param title: The title of this BookFields.
        :type title: str
        """

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this BookFields.


        :return: The author of this BookFields.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this BookFields.


        :param author: The author of this BookFields.
        :type author: str
        """

        self._author = author

    @property
    def year(self) -> int:
        """Gets the year of this BookFields.


        :return: The year of this BookFields.
        :rtype: int
        """
        return self._year

    @year.setter
    def year(self, year: int):
        """Sets the year of this BookFields.


        :param year: The year of this BookFields.
        :type year: int
        """

        self._year = year