# week7/flask-server/openapi_server/models/login_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class LoginResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, access_token=None, token_type=None):  # noqa: E501
        """LoginResponse - a model defined in OpenAPI

        :param access_token: The access_token of this LoginResponse.  # noqa: E501
        :type access_token: str
        :param token_type: The token_type of this LoginResponse.  # noqa: E501
        :type token_type: str
        """
        self.openapi_types = {
            'access_token': str,
            'token_type': str
        }

        self.attribute_map = {
            'access_token': 'access_token',
            'token_type': 'token_type'
        }

        self._access_token = access_token
        self._token_type = token_type

    @classmethod
    def from_dict(cls, dikt) -> 'LoginResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The LoginResponse of this LoginResponse.  # noqa: E501
        :rtype: LoginResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def access_token(self) -> str:
        """Gets the access_token of this LoginResponse.

        JWT access token  # noqa: E501

        :return: The access_token of this LoginResponse.
        :rtype: str
        """
        return self._access_token

    @access_token.setter
    def access_token(self, access_token: str):
        """Sets the access_token of this LoginResponse.

        JWT access token  # noqa: E501

        :param access_token: The access_token of this LoginResponse.
        :type access_token: str
        """
        if access_token is None:
            raise ValueError("Invalid value for `access_token`, must not be `None`")  # noqa: E501

        self._access_token = access_token

    @property
    def token_type(self) -> str:
        """Gets the token_type of this LoginResponse.


        :return: The token_type of this LoginResponse.
        :rtype: str
        """
        return self._token_type

    @token_type.setter
    def token_type(self, token_type: str):
        """Sets the token_type of this LoginResponse.


        :param token_type: The token_type of this LoginResponse.
        :type token_type: str
        """
        if token_type is None:
            raise ValueError("Invalid value for `token_type`, must not be `None`")  # noqa: E501

        self._token_type = token_type