# week7/flask-server/openapi_server/models/paginated_books_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server.models.book import Book
from openapi_server.models.pagination_metadata import PaginationMetadata
from openapi_server import util

from openapi_server.models.book import Book  # noqa: E501
from openapi_server.models.pagination_metadata import PaginationMetadata  # noqa: E501

class PaginatedBooksResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, data=None, metadata=None):  # noqa: E501
        """PaginatedBooksResponse - a model defined in OpenAPI

        :param data: The data of this PaginatedBooksResponse.  # noqa: E501
        :type data: List[Book]
        :param metadata: The metadata of this PaginatedBooksResponse.  # noqa: E501
        :type metadata: PaginationMetadata
        """
        self.openapi_types = {
            'data': List[Book],
            'metadata': PaginationMetadata
        }

        self.attribute_map = {
            'data': 'data',
            'metadata': 'metadata'
        }

        self._data = data
        self._metadata = metadata

    @classmethod
    def from_dict(cls, dikt) -> 'PaginatedBooksResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PaginatedBooksResponse of this PaginatedBooksResponse.  # noqa: E501
        :rtype: PaginatedBooksResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> List[Book]:
        """Gets the data of this PaginatedBooksResponse.


        :return: The data of this PaginatedBooksResponse.
        :rtype: List[Book]
        """
        return self._data

    @data.setter
    def data(self, data: List[Book]):
        """Sets the data of this PaginatedBooksResponse.


        :param data: The data of this PaginatedBooksResponse.
        :type data: List[Book]
        """
        if data is None:
            raise ValueError("Invalid value for `data`, must not be `None`")  # noqa: E501

        self._data = data

    @property
    def metadata(self) -> PaginationMetadata:
        """Gets the metadata of this PaginatedBooksResponse.


        :return: The metadata of this PaginatedBooksResponse.
        :rtype: PaginationMetadata
        """
        return self._metadata

    @metadata.setter
    def metadata(self, metadata: PaginationMetadata):
        """Sets the metadata of this PaginatedBooksResponse.


        :param metadata: The metadata of this PaginatedBooksResponse.
        :type metadata: PaginationMetadata
        """
        if metadata is None:
            raise ValueError("Invalid value for `metadata`, must not be `None`")  # noqa: E501

        self._metadata = metadata